Wasm Exploration.
I realized the WASM client wasn't using our protobuf types, and this is
the exploration of what it would look like to change that. It actually
seems like it's doable, but a lot of the architecture in the core crate
would have to change. So much, in fact, that it seems like we ought to
use a different approach.

This serves as a good reference point, just the same.
