+++ Cargo.toml
@@ -38,6 +38,7 @@ reqwest = { version = "0.12", features = ["json"] }
 rust-embed = { version = "8", features = ["axum-ex"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+serde-wasm-bindgen = "0.6.5"
 service-kit = { version = "0.1.0", path = "./service-kit" }
 service-kit-core = { version = "0.1.0", path = "./service-kit-core" }
 service-kit-web = { version = "0.1.0", path = "./service-kit-web" }
