+++ service-kit-core/Cargo.toml
@@ -6,39 +6,42 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-axum = { workspace = true }
-axum-extra = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true }
 color-eyre = { workspace = true }
 config = { workspace = true }
-crossterm = { workspace = true }
 dialoguer = { workspace = true }
 dirs = { workspace = true }
 futures = { workspace = true }
 indicatif = { workspace = true }
 mime_guess = { workspace = true }
+prost = { workspace = true }
 rand = { workspace = true }
-ratatui = { workspace = true }
 reqwest = { workspace = true }
-rust-embed = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-service-manager = { workspace = true }
-sqlx = { workspace = true }
+serde-wasm-bindgen = { workspace = true }
 strum = { workspace = true }
-prost = { workspace = true }
 thiserror = { workspace = true }
-tokio = { workspace = true }
-tokio-util = { workspace = true }
 tower-http = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
-tui-textarea = { workspace = true }
 wasm-bindgen = { workspace = true }
 wasm-bindgen-futures = { workspace = true }
 
-[dev-dependencies]
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+axum = { workspace = true }
+axum-extra = { workspace = true }
+crossterm = { workspace = true }
+ratatui = { workspace = true }
+rust-embed = { workspace = true }
+service-manager = { workspace = true }
+sqlx = { workspace = true }
+tui-textarea = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 tempfile = { workspace = true }
 
 [build-dependencies]
