+++ service-kit-core/src/context.rs
@@ -1,13 +1,13 @@
 #[derive(Clone, Debug)]
 pub struct WebContext {
     pub settings: crate::settings::Settings,
-    pub network: crate::settings::NetworkSettings,
+    pub network: crate::settings::network_settings::NetworkSettings,
     pub storage: crate::storage::StorageCollection,
 }
 
 impl WebContext {
     pub async fn new(
-        network: crate::settings::NetworkSettings,
+        network: crate::settings::network_settings::NetworkSettings,
         settings: crate::settings::Settings,
     ) -> crate::Result<Self> {
         Ok(Self {
@@ -17,6 +17,7 @@ impl WebContext {
         })
     }
 
+    #[cfg(not(target_arch = "wasm32"))]
     pub async fn listener(&self) -> crate::Result<tokio::net::TcpListener> {
         self.network.listener().await
     }
