+++ service-kit-core/src/lib.rs
@@ -1,14 +1,21 @@
 mod client;
+#[cfg(not(target_arch = "wasm32"))]
 mod context;
 mod errors;
+#[cfg(not(target_arch = "wasm32"))]
 mod server;
+#[cfg(not(target_arch = "wasm32"))]
 mod service;
 mod settings;
+#[cfg(not(target_arch = "wasm32"))]
 mod storage;
 mod telemetry;
+
+#[cfg(not(target_arch = "wasm32"))]
 mod tui;
 
 pub use client::WebClient;
+#[cfg(not(target_arch = "wasm32"))]
 pub use context::WebContext;
 pub use errors::Error;
