+++ service-kit-core/src/settings.rs
@@ -1,13 +1,13 @@
-use std::path::PathBuf;
-use std::str::FromStr;
+pub mod network_settings;
+
+mod client_settings;
+mod server_settings;
+mod service_settings;
 
 use clap::{Parser, ValueEnum};
 use config::{Config, ValueKind};
 use serde::Deserialize;
-use service_manager::ServiceManagerKind;
-use strum::{EnumString, VariantNames};
-
-use crate::Error;
+use std::path::PathBuf;
 
 /// A CLI application that helps do non-standard AzerothCore db tasks
 #[derive(Clone, Debug, Parser)]
@@ -43,217 +43,9 @@ pub struct GlobalOpts {
 pub enum Command {
     Debug,
     Tui,
-    Server(Server),
-    Client(Client),
-    Service(Service),
-}
-
-#[derive(Clone, Debug, Parser)]
-#[clap(rename_all = "kebab-case")]
-pub struct Server {
-    /// The mode to run the server in.
-    #[clap(subcommand)]
-    pub mode: Option<ServerMode>,
-    /// The settings for the server.
-    #[clap(flatten)]
-    pub settings: NetworkSettings,
-}
-
-#[derive(Clone, Debug, Parser, Default, EnumString, VariantNames)]
-#[clap(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab-case")]
-pub enum ServerMode {
-    /// Run the server in full mode.
-    #[default]
-    Full,
-    /// Run the server in a web mode.
-    Web,
-    /// Run the server in an api mode.
-    Api,
-}
-
-impl ServerMode {
-    pub fn options() -> &'static [&'static str] {
-        Self::VARIANTS
-    }
-
-    pub async fn exec(&self, config: crate::WebContext) -> crate::Result<()> {
-        Ok(match self {
-            ServerMode::Full => crate::server::init(config).await?,
-            ServerMode::Web => crate::server::web::init(config).await?,
-            ServerMode::Api => crate::server::api::init(config).await?,
-        })
-    }
-
-    pub fn select() -> crate::Result<Self> {
-        let options = Self::options();
-        let result = dialoguer::Select::with_theme(&dialoguer::theme::ColorfulTheme::default())
-            .with_prompt("Select a server mode")
-            .default(0)
-            .items(options)
-            .interact()
-            .expect("Unable to select server mode");
-
-        Ok(Self::from_str(options[result])?)
-    }
-}
-
-#[derive(Clone, Debug, Parser)]
-#[clap(rename_all = "kebab-case")]
-pub struct Client {
-    /// Tell the client what resource to connect to.
-    #[clap(subcommand)]
-    pub resource: Option<ClientResource>,
-    /// The settings for the client.
-    #[clap(flatten)]
-    pub settings: NetworkSettings,
-}
-
-#[derive(Clone, Debug, Parser, EnumString, VariantNames)]
-#[clap(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab-case")]
-pub enum ClientResource {
-    /// The health check api.
-    Health,
-}
-
-impl ClientResource {
-    pub fn options() -> &'static [&'static str] {
-        Self::VARIANTS
-    }
-
-    pub async fn exec(&self, config: NetworkSettings) -> crate::Result<()> {
-        Ok(match self {
-            ClientResource::Health => crate::client::health(config).await?,
-        })
-    }
-
-    pub fn select() -> crate::Result<Self> {
-        let options = Self::options();
-        let result = dialoguer::Select::with_theme(&dialoguer::theme::ColorfulTheme::default())
-            .with_prompt("Select a client resource")
-            .default(0)
-            .items(options)
-            .interact()
-            .expect("Unable to select client resource");
-
-        Ok(Self::from_str(options[result])?)
-    }
-}
-
-#[derive(Clone, Debug, Default, Parser)]
-pub struct NetworkSettings {
-    /// The host to connect to.
-    #[clap(long, default_value = "localhost")]
-    pub host: String,
-
-    /// The port to connect to.
-    #[clap(long, default_value = "8080")]
-    pub port: u16,
-}
-
-impl NetworkSettings {
-    pub fn address(&self) -> String {
-        format!("{}:{}", self.host, self.port)
-    }
-
-    pub async fn listener(&self) -> crate::Result<tokio::net::TcpListener> {
-        tokio::net::TcpListener::bind(self.address())
-            .await
-            .map_err(Error::ListenerInitFailure)
-    }
-}
-
-#[derive(Clone, Debug, Parser)]
-#[clap(rename_all = "kebab-case")]
-pub struct Service {
-    /// Control the service itself.
-    #[clap(subcommand)]
-    pub operation: Option<ServiceOperation>,
-    #[clap(flatten)]
-    pub settings: ServiceSettings,
-}
-
-#[derive(Clone, Debug, Parser)]
-pub struct ServiceSettings {
-    /// The service label to use. Defaults to the app name.
-    #[clap(long)]
-    pub service_label: Option<String>,
-    /// The kind of service manager to use. Defaults to system native.
-    #[clap(long, value_enum)]
-    pub service_manager: Option<ServiceManagerKind>,
-    /// Install system-wide. If not set, attempts to install for the current user.
-    #[clap(long)]
-    pub system: bool,
-}
-
-#[derive(Clone, Debug, Parser, EnumString, VariantNames)]
-#[clap(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab-case")]
-pub enum ServiceOperation {
-    Install,
-    Uninstall,
-    Start,
-    Stop,
-}
-
-impl ServiceOperation {
-    pub fn options() -> &'static [&'static str] {
-        Self::VARIANTS
-    }
-
-    pub fn select() -> crate::Result<Self> {
-        let options = Self::options();
-        let result = dialoguer::Select::with_theme(&dialoguer::theme::ColorfulTheme::default())
-            .with_prompt("Select a service operation")
-            .default(0)
-            .items(options)
-            .interact()
-            .expect("Unable to select service operation");
-
-        Ok(Self::from_str(options[result])?)
-    }
-
-    pub async fn exec(&self, cli: Cli, settings: ServiceSettings) -> crate::Result<()> {
-        let program = std::env::current_exe()?;
-        let args: Vec<std::ffi::OsString> = vec![
-            "-a".into(),
-            cli.global.app_name.clone().into(),
-            "server".into(),
-            "api".into(),
-        ];
-        let service = crate::service::Service::init(
-            settings
-                .service_label
-                .clone()
-                .unwrap_or_else(|| format!("local.service.{}", cli.global.app_name))
-                .parse()?,
-        )?;
-
-        match self {
-            ServiceOperation::Install => service.install(program, args)?,
-            ServiceOperation::Start => service.start()?,
-            ServiceOperation::Stop => service.stop()?,
-            ServiceOperation::Uninstall => service.uninstall()?,
-        }
-
-        Ok(())
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum NumberOrString {
-    String(String),
-    Number(u32),
-}
-
-impl std::str::FromStr for NumberOrString {
-    type Err = &'static str; // The actual type doesn't matter since we never error, but it must implement `Display`
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(s.parse::<u32>()
-            .map(NumberOrString::Number)
-            .unwrap_or_else(|_| NumberOrString::String(s.to_string())))
-    }
+    Server(server_settings::Server),
+    Client(client_settings::Client),
+    Service(service_settings::Service),
 }
 
 #[derive(Clone, Copy, Debug, Default, Deserialize, strum::Display, ValueEnum)]
@@ -375,6 +167,7 @@ impl Settings {
             Command::Tui => {
                 tracing::info!("Starting TUI");
 
+                #[cfg(not(target_arch = "wasm32"))]
                 crate::tui::init().await?;
             }
             Command::Server(server_details) => {
@@ -388,7 +181,7 @@ impl Settings {
                     None => {
                         tracing::info!("No server mode specified, prompting");
 
-                        ServerMode::select()?.exec(context).await?;
+                        server_settings::ServerMode::select()?.exec(context).await?;
                     }
                 }
             }
@@ -400,7 +193,7 @@ impl Settings {
                     None => {
                         tracing::info!("No client resource specified, prompting");
 
-                        ClientResource::select()?
+                        client_settings::ClientResource::select()?
                             .exec(client_details.settings)
                             .await?;
                     }
@@ -418,7 +211,7 @@ impl Settings {
                     None => {
                         tracing::info!("No service operation specified, prompting");
 
-                        ServiceOperation::select()?
+                        service_settings::ServiceOperation::select()?
                             .exec(self.cli.clone(), service_details.settings)
                             .await?;
                     }
