+++ service-kit/Cargo.toml
@@ -6,33 +6,4 @@ edition = "2021"
 [dependencies]
 service-kit-core = { workspace = true }
 service-kit-web = { workspace = true }
-axum = { workspace = true }
-axum-extra = { workspace = true }
-chrono = { workspace = true }
-clap = { workspace = true }
-color-eyre = { workspace = true }
-config = { workspace = true }
-crossterm = { workspace = true }
-dialoguer = { workspace = true }
-dirs = { workspace = true }
-futures = { workspace = true }
-indicatif = { workspace = true }
-mime_guess = { workspace = true }
-rand = { workspace = true }
-ratatui = { workspace = true }
-reqwest = { workspace = true }
-rust-embed = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-service-manager = { workspace = true }
-sqlx = { workspace = true }
-strum = { workspace = true }
-thiserror = { workspace = true }
 tokio = { workspace = true }
-tokio-util = { workspace = true }
-tower-http = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-tui-textarea = { workspace = true }
-wasm-bindgen = { workspace = true }
-wasm-bindgen-futures = { workspace = true }
